Fold useKeys into useSoftkey to prevent it from being used directly
