+++ src/components/Softkey.js
@@ -1,6 +1,6 @@
 import { h } from 'preact'
 import { memo } from 'preact/compat'
-import { useKeys } from 'hooks'
+import { useEffect, useRef } from 'preact/hooks'
 
 const SoftkeyBar = memo(({ left, center, right }) => {
   return (
@@ -14,17 +14,18 @@ const SoftkeyBar = memo(({ left, center, right }) => {
 
 export const Softkey = ({
   left,
-  onKeyLeft,
   center,
-  onKeyCenter,
   right,
+  onKeyLeft,
+  onKeyCenter,
   onKeyRight,
   onKeyArrowDown,
   onKeyArrowUp,
   onKeyArrowLeft,
   onKeyArrowRight
 }) => {
-  useKeys({
+  const handlersRef = useRef()
+  handlersRef.current = {
     SoftLeft: onKeyLeft,
     l: onKeyLeft, // for testing purposes
     Enter: onKeyCenter,
@@ -34,7 +35,20 @@ export const Softkey = ({
     ArrowUp: onKeyArrowUp,
     ArrowLeft: onKeyArrowLeft,
     ArrowRight: onKeyArrowRight
-  })
+  }
+  const onKeyDown = (e) => {
+    const key = e.key.toString()
+    if (handlersRef.current[key]) {
+      handlersRef.current[key](e)
+      e.stopPropagation()
+      e.preventDefault()
+    }
+  }
+
+  useEffect(() => {
+    document.addEventListener('keydown', onKeyDown)
+    return () => document.removeEventListener('keydown', onKeyDown)
+  }, [])
 
   return <SoftkeyBar left={left} center={center} right={right} />
 }
