+++ src/hooks/index.js
@@ -2,7 +2,6 @@ export * from './useArticle'
 export * from './useHistoryState'
 export * from './useI18n'
 export * from './useSoftkey'
-export * from './useKeys'
 export * from './useLanguage'
 export * from './useNavigation'
 export * from './useScroll'
