--- src/hooks/useKeys.js
@@ -1,19 +1,0 @@
-import { useEffect, useRef } from 'preact/hooks'
-
-export const useKeys = (handlers) => {
-  const handlersRef = useRef()
-  handlersRef.current = handlers
-  const onKeyDown = (e) => {
-    const key = e.key.toString()
-    if (handlersRef.current[key]) {
-      handlersRef.current[key](e)
-      e.stopPropagation()
-      e.preventDefault()
-    }
-  }
-
-  useEffect(() => {
-    document.addEventListener('keydown', onKeyDown)
-    return () => document.removeEventListener('keydown', onKeyDown)
-  }, [])
-}
